# Only compute nanoTime once for a root span.