Currently we compute current `nanoTime` twice for a root span unnecessarily. `nanoTime` calls into the OS so it's not the fastest API out there and this does end up showing up on flame graphs. It's not so difficult to avoid this.

After
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
FillSpanBenchmark.setFourAttributes                               thrpt   60  2164.882 ± 28.919  ops/ms
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate                thrpt   60   813.621 ± 10.910  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate.norm           thrpt   60   592.000 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space       thrpt   60   813.659 ± 24.409  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space.norm  thrpt   60   591.916 ± 15.074    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen          thrpt   60     0.002 ±  0.001  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen.norm     thrpt   60     0.001 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.count                     thrpt   60   457.000           counts
FillSpanBenchmark.setFourAttributes:·gc.time                      thrpt   60   174.000               ms
```

Before
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
FillSpanBenchmark.setFourAttributes                               thrpt   60  2029.646 ± 25.532  ops/ms
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate                thrpt   60   763.105 ±  9.714  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate.norm           thrpt   60   592.000 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space       thrpt   60   762.296 ± 25.178  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space.norm  thrpt   60   591.415 ± 18.165    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen          thrpt   60     0.002 ±  0.001  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen.norm     thrpt   60     0.002 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.count                     thrpt   60   427.000           counts
FillSpanBenchmark.setFourAttributes:·gc.time                      thrpt   60   164.000               ms
```